Support older versions of GTK+ than the 2.14
gtk_dialog_get_action_area () and gtk_dialog_get_content_area ()
apperead in GTK+-2.14.

For GtkSourceView, the 2.4 release is required because of
gtk_source_language_manager_guess_language (). It is possible to support
older versions of GtkSourceView by checking with CMake if the version is
older, and then call an other function in that case. But I prefer not to do
that because GtkSourceView 2.4 was released one year ago, so it's not
worth it.