Show output command of the compilation progressively
With the use of GIOChannels, the output of the compilation command
(latex or pdflatex) is showed exactly the same as if the command were
executed in a terminal.
Now all the functions for the actions which appear in the action history
and for which there is a log message are moved to "actions.c" and
"actions.h".