Save and load preferences
Now the prefereneces related functions are in prefs.c. The rc file is
store in $HOME/.config/latexila/latexilarc. There is a structure
preferences_t wich hold all the preferences, it is easier to store and
load values than the GKeyFile. The rc file is saved only at exit.

Now all the dimensions of the program are stored: width and height of
the window, if the window is maximized and the positions of the panes.