Unmaximization: best method
When saving preferences, if the window is maximized, the sizes of the
window are reduced by 100, else making window non-maximized the next
time will have no effect.

There was previously a problem with this method because the left/top
widgets was packed into the panes with "gtk_paned_pack1 (paned, widget,
TRUE, TRUE)", so the position of the panes increased... Now the widgets
are packed into the panes with the gtk_paned_add* functions.

If needed, the previous method is in the previous commit ;)