Other method for the log zone
For each action (compile, view or convert a document) there is a new
entry in the list, which contain now only 2 things: the title, and a
pointer to a GtkTextBuffer, which contain all the details (command
output, etc). So when we select an other entry in the list, we set the
right GtkTextBuffer to the GtkTextView. When we do an new action, a new
GtkTextBuffer is created and filled.