Sensitivity of the actions when an action is running
New structure action_t in main.h which contain all the actions used
later in the program (now only for setting the sensitivity).

In cb_watch_output_command (), there is a new static variable
"nb_channels_active" which is set to 2, because in compile_document ()
and in convert_document (), cb_watch_output_command is used for the
output and the error output. When there is a G_IO_HUP,
nb_channels_active is decremented, and when the value is 0, we call
command_running_finished () and we reset the static variables.
If we don't do that, command_running_finished is called two times, and
the sensitivity of the actions is set to true whereas the action is
still running.