File browser: sort the files in alphabetical order
The strcmp function is used. It's not perfect but it works :)
The file names are stored in two GList's: one for files, and the other for
directories. Then we call g_list_sort () to sort the lists in
alphabetical order. Then we traverse the two lists for adding all the
files in the list store.