Remember the files opened at exit and reopen them at starting
In callbacks.c there is a GList (list_opened_docs) which is filled in
the close_document () function, only when save_list_opened_docs is TRUE.
In cb_quit () we set save_list_opened_docs to TRUE, and close_document
is called inderectly when we call close_all ().
We use the g_key_file_set_string_list (), which take an array in
argument, not a GList, so we must convert the GList into an array, which
is done in cb_quit before calling save_preferences.