Some various improvements
First, for the entries of go to line and find, when there is an error
the background and text colors change. There is an error in the go to
line if the line specified is to high. There is an error in find when
the phrase is not found.

The main() function was too long, so it is splitted into some init_*
functions.

Bug correction: there was a segfault if we ran latexila like this:
$ latexila --new-document /a/file
(strcmp() can not compare a NULL value)

Maybe other little things without importance ;)