Side pane: remember the current page of the notebook
The page number of the notebook of the side pane is saved in the
preferences.

For the 0.1 version there is only two pages in the notebook (symbols and
file browser) and I think it's better than a ComboBox, but maybe in the
future it will be replaced by a ComboBox with a close button.