Output filter: extract info and show stats
The output filter is now almost the same as in Kile. The output of the
compilation is filtered so only usefull messages are showed. A message
can be an error, a warning, a badbox or an other thing. Later, this
messages will be displayed in different colors. Now, to distinguish the
messages, a symbol between [ ] is showed ("[x]" for an error, "[+]" for
a warning, etc).

The informations extracted are the files that TeX is processing, the
line and obviously the message.

There are statistics about how many errors, warnings and badboxes
occured during the compilation.