Log output: jump to lines and files
We can now click on a message and go to the file and the line specified.
Certain row can not be selected, only those with a real filename. If the
line number is not specified, simply go to the file (but not change the
cursor position).

There are also some fixes for memory leaks, that's why a lot of files
are modified ;)