Tooltip on document tabs: filename
The full path of the filename is set as a tooltip of the label in
document tabs. The only two places where we set it, is where a new
document is created: in create_document_in_new_tab(). And when the
path of the document change: in save_as_dialog().

There is a new string in the document_t structure: basename. Because we
often need it.