Log output: background colors for selected rows
If the user click on a row, output_row_selection_func() will be
called, but no row will be selected (the function return FALSE
everytime). Instead, if the row can be "selected" (if the message
type is an error, a warning or a badbox, and if the filename is not
empty), the background and foreground colors are inverted.

The GtkTreePath of the row selected is stored as a data in the
GtkTreeModel (with g_object_set_data()).