Sensitivity of previous/next latex messages actions
For each GtkTreeModel which contain errors or warnings or badboxes,
there is an index which contains the rows numbers for each category of
messages. The index is a messages_index_t structure. This structure is
created when print_output_stats() is called (because we must know the
numbers of errors/warnings/... to allocate memory). And the index is
filled when print_output_exit_code() is called.

So with the index it's easier and faster for setting the sensitivities
of the actions.