Create a backup copy of files before saving
There is a new variable in document_t: backup_made. backup_made is FALSE
when a document is created or opened. When a document is saved, and if
the option in the preferences is enabled (by default, yes), before
erasing the file, we rename the file with a name like "file~". When we
save as a document, backup_made is reset to FALSE.