Change description: "Integrated LaTeX Environment" -> "LaTeX editor"
There are some stuff that usually an IDE can do but is out of scope for
LaTeXila, for example support for Git (doing all git operations within
the IDE).

There is one thing that LaTeXila can do and is within the scope of an
IDE (a bit beyond a simple text editor), which is: compiling the
documents. But that's the only feature IDE-ish, and it is easy *not* to
use it.

Well, there is also project management which is IDE-ish, but managing a
project in LaTeXila is so simple, it is a really small feature (but
useful). It's not like a programming project like in Eclipse, creating a
jar etc.

LaTeXila is still a lightweight application for doing LaTeX, so "LaTeX
editor" better fits it, and has the advantage that it is shorter.

The AppData description has been improved as well (and copied in the
*.doap file).