Apply GtkSourceView coding style to *.c files
The script apply-gsv-coding-style-no-backup.sh from GtkSourceView has
been run on all the *.c files in gnome-latex. The script uses
uncrustify. The uncrustify config file is maybe not perfect, if someone
sees a problem, the uncrustify config file can be improved and then
re-run on all the *.c files.

The *.h files have not been processed, because the script cannot
correctly format function prototypes. The headers will either be
manually adapted, or another script can be written.