Release 3.28.0
A bit late, but I prefer to release often (6 months is already a long
time), to follow the "release early, release often" principle.

What was missing for the GNOME 3.28.0 date is to import the settings
from latexila, which is now done.

Another thing missing, but that doesn't depend on me, is a new icon.