build: add option to disable dconf migration support
dconf is not available on Mac OS X or Windows.

It is also useful on Linux because dconf has an unstable API: if dconf
breaks its API in the future and gnome-latex is unmaintained, it will
still be possible to compile gnome-latex by disabling the dconf
migration support.

Fixes https://gitlab.gnome.org/GNOME/gnome-latex/issues/56