Revert "LaTeX commands: Avoid superfluous and fatal cast to Gtk.Activatable"
This reverts commit c56caaa516c8371a6ac179661978b720a731fd65.

For the 3.28.x versions I must compile with vala 0.40 to avoid all the
crashes:
https://gitlab.gnome.org/GNOME/gnome-latex/issues/57
https://gitlab.gnome.org/GNOME/gnome-latex/issues/59
(and there are probably more).

With this revert, if I compile gnome-latex 3.28 with vala git master,
I'll get a compilation error:
error: type `Activatable' does not support floating references

In the tarballs the generated C code is shipped, so there shouldn't be
any problem, except if the distribution prefers to re-generate the C
code (I need to add a warning in the NEWS file).