Revert "LaTeX commands: Avoid superfluous and fatal cast to Gtk.Activatable"
This reverts commit 02f35b9dc60eab0cf741e44740e12c91daf9ea8e.

I've discovered more crashes when gnome-latex is compiled with valac git
master (what will become vala 0.42). I don't plan to work on gnome-latex
anytime soon, so in the meantime gnome-latex will need to be compiled
with vala 0.40. With this revert the nice thing is that gnome-latex
cannot be compiled with vala git master, but can be compiled with vala
0.40.