flatpak: Add flatpak manifest for gnome-latex
Adds a flatpak manifest for gnome-latex, which seems to function
properly. This uses the new org.freedesktop.Sdk.Extension.texlive
extension in flathub [0] to solve the issue of how to handle including
texlive.

Unfortunately, gnome-latex does not compile against the latest version
of vala, which is used in the SDK, so I had to bundle vala with the app.
This also required a patch to stop building valadoc, and thus remove a
dependency on graphviz.

As the extension used is for the SDK, rather than the platform, this app
has to use org.gnome.Sdk as the runtime, otherwise the extension is not
loaded. Additionally $PATH is modified specifically for x86_64-linux,
however as the texlive extension is only compiled for x86_64 this app
wouldn't be able to run using this extension on other architectures anyway.

[0]: https://github.com/flathub/org.freedesktop.Sdk.Extension.texlive