App: update copyright in about dialog for 2019 (10 years!)
In July 24 of this year, it'll be exactly 10 years of
LaTeXila/GNOME LaTeX!

From the reviews on gnome-software, I know that at least one person has
used this app to write his PhD thesis. By going to the GUADEC conference
I also know that one person use or used this app almost every day. So I
think that's a good achievement. It permitted me to learn a lot, to
practice programming during the summer when I was a student, and then
contributing down the stack (especially to the GtkSourceView library).